Add utility class to create symbols
